# deepexi-dd-middle-order-qa/deepexi-gxs-middle-order-api/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "swagger导入dapi demo",
    "version" : "v1",
    "title" : "swagger导入dapi demo",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "deepexi",
      "url" : "www.deepexi.com",
      "email" : "deepexi@deepexi.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "petstore.swagger.wordnik.com",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/middle/v1/saleOrderInfos/{id}" : {
      "get" : {
        "tags" : [ "销售订单表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PayloadSaleOrderInfoResponseVO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售订单表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SaleOrderInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PayloadBoolean"
            }
          }
        }
      }
    },
...
    "/orderConsigneeInfos/{id}" : {
      "get" : {
        "tags" : [ "收货地址信息表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderConsigneeInfoResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "收货地址信息表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderConsigneeInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/orderCouponInfos/{id}" : {
      "get" : {
        "tags" : [ "优惠信息表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderCouponInfoResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "优惠信息表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderCouponInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/orderExpenseInfos/{id}" : {
      "get" : {
        "tags" : [ "销售订单的费用信息表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderExpenseInfoResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售订单的费用信息表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderExpenseInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/orderInvoiceInfos/{id}" : {
      "get" : {
        "tags" : [ "订单的开票信息表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderInvoiceInfoResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "订单的开票信息表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderInvoiceInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/orderPromotionInfos/{id}" : {
      "get" : {
        "tags" : [ "销售订单的促销信息表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderPromotionInfoResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售订单的促销信息表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderPromotionInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/orderReturnInfos/{id}" : {
      "get" : {
        "tags" : [ "销售订单退货单信息表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderReturnInfoResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售订单退货单信息表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderReturnInfoRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/orderReturnItems/{id}" : {
      "get" : {
        "tags" : [ "销售订单退货明细表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderReturnItemResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售订单退货明细表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderReturnItemRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/saleOrderItems/{id}" : {
      "get" : {
        "tags" : [ "销售订单明细表管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SaleOrderItemResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售订单明细表管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SaleOrderItemRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
...
    "/saleOutTasks/{id}" : {
      "get" : {
        "tags" : [ "销售出库单（出库任务单）管理" ],
        "operationId" : "selectById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SaleOutTaskResponseDTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "销售出库单（出库任务单）管理" ],
        "operationId" : "updateById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SaleOutTaskRequestDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "OrderConsigneeInfoRequestDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "tenantId" : {
          "type" : "string",
          "description" : "租户ID"
        },
        "appId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "应用ID"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "版本号"
        },
        "remark" : {
          "type" : "string",
          "description" : "备注"
        },
        "deleted" : {
          "type" : "boolean",
          "description" : "0：未逻辑删除状态。1：删除"
        },
        "createdTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updatedTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        },
        "saleOrderId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "订单ID"
        },
        "saleOrderCode" : {
          "type" : "string",
          "description" : "订单编号"
        },
        "provinceCode" : {
          "type" : "string",
          "description" : "省份编号"
        },
        "provinceName" : {
          "type" : "string",
          "description" : "省份名称"
        },
        "cityCode" : {
          "type" : "string",
          "description" : "城市编号"
        },
        "cityName" : {
          "type" : "string",
          "description" : "城市名称"
        },
        "areaCode" : {
          "type" : "string",
          "description" : "区域编号"
        },
        "areaName" : {
          "type" : "string",
          "description" : "区域名称"
        },
        "streetCode" : {
          "type" : "string",
          "description" : "街道编号"
        },
        "streetName" : {
          "type" : "string",
          "description" : "街道名称"
        },
        "consignee" : {
          "type" : "string",
          "description" : "收货人"
        },
        "mobile" : {
          "type" : "string",
          "description" : "手机号码"
        }
      },
      "description" : "OrderConsigneeInfoRequestDTO"
    },
    "OrderConsigneeInfoResponseDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "tenantId" : {
          "type" : "string",
          "description" : "租户ID"
        },
        "appId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "应用ID"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "版本号"
        },
        "remark" : {
          "type" : "string",
          "description" : "备注"
        },
        "deleted" : {
          "type" : "boolean",
          "description" : "0：未逻辑删除状态。1：删除"
        },
        "createdTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updatedTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        },
        "saleOrderId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "订单ID"
        },
        "saleOrderCode" : {
          "type" : "string",
          "description" : "订单编号"
        },
        "provinceCode" : {
          "type" : "string",
          "description" : "省份编号"
        },
        "provinceName" : {
          "type" : "string",
          "description" : "省份名称"
        },
        "cityCode" : {
          "type" : "string",
          "description" : "城市编号"
        },
        "cityName" : {
          "type" : "string",
          "description" : "城市名称"
        },
        "areaCode" : {
          "type" : "string",
          "description" : "区域编号"
        },
        "areaName" : {
          "type" : "string",
          "description" : "区域名称"
        },
        "streetCode" : {
          "type" : "string",
          "description" : "街道编号"
        },
        "streetName" : {
          "type" : "string",
          "description" : "街道名称"
        },
        "consignee" : {
          "type" : "string",
          "description" : "收货人"
        },
        "mobile" : {
          "type" : "string",
          "description" : "手机号码"
        }
      },
      "description" : "OrderConsigneeInfoResponseDTO"
    },
...
    "OrderCouponInfoRequestDTO" : {
      "type" : "object",
      "required" : [ "tenantId" ],
      "properties" : {
        "tenantId" : {
          "type" : "string",
          "description" : "租户id"
        },
        "appId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "应用id"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "ID"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "版本号"
        },
        "remark" : {
          "type" : "string",
          "description" : "备注"
        },
        "deleted" : {
          "type" : "boolean",
          "description" : "0：未逻辑删除状态。1:删除"
        },
        "createdTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updatedTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        },
        "couponId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "优惠券ID"
        },
        "couponCode" : {
          "type" : "string",
          "description" : "优惠券编码"
        },
        "couponType" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "优惠券类型"
        },
        "couponName" : {
          "type" : "string",
          "description" : "优惠券名称"
        },
        "couponQuntity" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "优惠券数量"
        },
        "faceValue" : {
          "type" : "number",
          "description" : "面值金额"
        },
        "couponRule" : {
          "type" : "string",
          "description" : "优惠规则"
        },
        "expiredTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "优惠券过期时间"
        },
        "skuIds" : {
          "type" : "string",
          "description" : "sku商品ids"
        },
        "skuNames" : {
          "type" : "string",
          "description" : "sku名称、规格"
        },
        "discountAmount" : {
          "type" : "number",
          "description" : "优惠金额"
        }
      },
      "description" : "优惠券对象"
    },
    "OrderCouponInfoResponseDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "ID"
        },
        "tenantId" : {
          "type" : "string",
          "description" : "租户ID"
        },
        "appId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "应用ID"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "版本号"
        },
        "remark" : {
          "type" : "string",
          "description" : "备注"
        },
        "deleted" : {
          "type" : "boolean",
          "description" : "0：未逻辑删除状态。1：删除"
        },
        "createdTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updatedTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        },
        "saleOrderId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "订单ID"
        },
        "saleOrderCode" : {
          "type" : "string",
          "description" : "订单编号"
        },
        "couponId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "优惠券ID"
        },
        "couponCode" : {
          "type" : "string",
          "description" : "优惠券编码"
        },
        "couponType" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "优惠券类型"
        },
        "couponName" : {
          "type" : "string",
          "description" : "优惠券名称"
        },
        "couponQuntity" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "优惠券数量"
        },
        "faceValue" : {
          "type" : "number",
          "description" : "面值金额"
        },
        "couponRule" : {
          "type" : "string",
          "description" : "优惠规则"
        },
        "expiredTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "优惠券过期时间"
        },
        "skuIds" : {
          "type" : "string",
          "description" : "sku商品ids"
        },
        "skuNames" : {
          "type" : "string",
          "description" : "sku名称、规格"
        },
        "discountAmount" : {
          "type" : "number",
          "description" : "优惠金额"
        }
      },
      "description" : "OrderCouponInfoResponseDTO"
    },
...
    "PayloadBoolean" : {
      "type" : "object",
      "properties" : {
        "payload" : {
          "type" : "boolean"
        },
        "code" : {
          "type" : "string"
        },
        "msg" : {
          "type" : "string"
        }
      }
    },
...
    "PayloadSaleOrderInfoResponseVO" : {
      "type" : "object",
      "properties" : {
        "payload" : {
          "$ref" : "#/definitions/SaleOrderInfoResponseVO"
        },
        "code" : {
          "type" : "string"
        },
        "msg" : {
          "type" : "string"
        }
      }
    },
    "SaleOrderInfoRequestDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "ID"
        },
        "tenantId" : {
          "type" : "string",
          "description" : "租户ID"
        },
        "appId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "APP标识"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "版本号"
        },
        "remark" : {
          "type" : "string",
          "description" : "备注"
        },
        "deleted" : {
          "type" : "boolean",
          "description" : "0：未逻辑删除状态。1:删除"
        },
        "createdTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updatedTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        },
        "code" : {
          "type" : "string",
          "description" : "订单编号"
        },
        "status" : {
          "type" : "string",
          "description" : "订单状态:draft 草稿,accept 待接单,deliver 待发货,receipt 待收货,examining 接单审核中,reject 接单驳回，payment 待付款，paid 收款待确认，over 交易完成，cancel 已取消"
        },
        "sellerId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "供货商ID"
        },
        "sellerName" : {
          "type" : "string",
          "description" : "供货商名称"
        },
        "buyerId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "客户ID"
        },
        "buyerType" : {
          "type" : "string",
          "description" : "下单类型:(agent:代客下单,mall:h5商城下单)"
        },
        "buyerName" : {
          "type" : "string",
          "description" : "客户名称"
        },
        "ticketType" : {
          "type" : "string",
          "description" : "单据类型(ordinary:普通销售单)"
        },
        "shippingType" : {
          "type" : "string",
          "description" : "送货方式(logistics:物流配送)"
        },
        "shipmentStatus" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "出库状态 0:未出库,1已出库"
        },
        "fromStorehouse" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "发货仓库"
        },
        "handler" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "经手人"
        },
        "ticketDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "单据日期"
        },
        "totalAmount" : {
          "type" : "number",
          "description" : "总商品金额"
        },
        "discountAmount" : {
          "type" : "number",
          "description" : "促销优惠金额"
        },
        "accrueAmount" : {
          "type" : "number",
          "description" : "应收金额"
        },
        "totalExpense" : {
          "type" : "number",
          "description" : "费用小计"
        },
        "createdBy" : {
          "type" : "string",
          "description" : "创建人"
        },
        "updatedBy" : {
          "type" : "string",
          "description" : "更新人"
        },
        "paymentType" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "支付类型:1线下支付,2信用支付"
        },
        "settlementStatus" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "结算状态 0:未结算,1:已结算,2:部分结算"
        },
        "paymentStatus" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "支付状态 0:未付款,1:已付款,2:部分付款"
        },
        "arriveDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "要求送达日期"
        },
        "isolationId" : {
          "type" : "string",
          "description" : "隔离标识"
        },
        "handlerName" : {
          "type" : "string",
          "description" : "经手人名称"
        },
        "department" : {
          "type" : "string",
          "description" : "部门名称"
        }
      },
      "description" : "SaleOrderInfoRequestDTO"
    },
...
  }
}